refactor(fs): pathFind.ts
Problem:
SystemUtilities is deprecated and redundant.

Solution:
Move remaining logic into new `pathFind.ts` module.
